requirements.txt
asgiref==3.3.4
boto3==1.17.74
botocore==1.20.74
dj-database-url==0.5.0
Django==3.2.3
django-localflavor==3.0.1
django-storages==1.11.1
geographiclib==1.50
geopy==2.0.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==7.1.2
python-dateutil==2.8.1
python-stdnum==1.16
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
psycopg2==2.8.6
